Update DB Layer to use Async (#302)
* Update DB Layer to use Async

* add async stuff, but not working. reorganize the config

* webserver starts up

* queries are running but coming back empty

* queries are working, but the return type is still wrong

* get and list both now work

* delete and insert working

* fix update

* most tests pass

* all tests passing

* fix linting issues

* all checks passing

* update the docs with a note about the lack of driver specification

* all checks passing locally

* code reorg

* add the explicit session begin

* move the mypy ignore into the source files directly

* fix tests

* add await to list_resource(Evaluation) in view.py
