+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     command: uvicorn --host 0.0.0.0 --port 8080 --reload fidesapi.main:app
     healthcheck:
       test: [ "CMD", "curl", "-f", "http://0.0.0.0:8080/health" ]
-      interval: 5s
+      interval: 15s
       timeout: 5s
       retries: 5
     depends_on:
@@ -26,7 +26,7 @@ services:
     image: postgres:12
     healthcheck:
       test: [ "CMD-SHELL", "pg_isready -U postgres" ]
-      interval: 5s
+      interval: 15s
       timeout: 5s
       retries: 5
     volumes:
