+++ fidesctl/setup.cfg
@@ -29,9 +29,3 @@ warn_untyped_fields = True
 
 [mypy-src.fidesctl._version]
 ignore_errors = True
-
-[mypy-src.fidesapi.crud]
-ignore_errors = True
-
-[mypy-src.fidesapi.sql_models]
-ignore_errors = True
