+++ fidesctl/src/fidesapi/database/database.py
@@ -1,5 +1,5 @@
 """
-Contains all of the logic for spinning up/tearing down the database.
+Contains all of the logic related to the database including connections, setup, teardown, etc.
 """
 import os
 
@@ -9,11 +9,12 @@ from alembic.migration import MigrationContext
 from loguru import logger as log
 from sqlalchemy_utils.functions import create_database, database_exists
 
-from fidesapi.errors import QueryError
+from fidesapi.utils.errors import QueryError
 from fidesapi.sql_models import sql_model_map, SqlAlchemyBase
-from fidesapi.crud import upsert_resources
 from fideslang import DEFAULT_TAXONOMY
 from fidesctl.core.utils import get_db_engine
+
+from .crud import upsert_resources
 
 
 def get_alembic_config(database_url: str) -> Config:
@@ -22,8 +23,8 @@ def get_alembic_config(database_url: str) -> Config:
     """
 
     migrations_dir = os.path.dirname(os.path.abspath(__file__))
-    directory = os.path.join(migrations_dir, "migrations")
-    config = Config(os.path.join(migrations_dir, "alembic.ini"))
+    directory = os.path.join(migrations_dir, "../migrations")
+    config = Config(os.path.join(migrations_dir, "../alembic.ini"))
     config.set_main_option("script_location", directory.replace("%", "%%"))
     config.set_main_option("sqlalchemy.url", database_url)
     return config
@@ -35,14 +36,14 @@ def upgrade_db(alembic_config: Config, revision: str = "head") -> None:
     command.upgrade(alembic_config, revision)
 
 
-def init_db(database_url: str) -> None:
+async def init_db(database_url: str) -> None:
     """
     Runs the migrations and creates all of the database objects.
     """
     log.info("Initializing database")
     alembic_config = get_alembic_config(database_url)
     upgrade_db(alembic_config)
-    load_default_taxonomy()
+    await load_default_taxonomy()
 
 
 def create_db_if_not_exists(database_url: str) -> None:
@@ -53,7 +54,7 @@ def create_db_if_not_exists(database_url: str) -> None:
         create_database(database_url)
 
 
-def load_default_taxonomy() -> None:
+async def load_default_taxonomy() -> None:
     "Upserts the default taxonomy into the database."
     log.info("UPSERTING the default fideslang taxonomy")
     for resource_type in list(DEFAULT_TAXONOMY.__fields_set__):
@@ -61,7 +62,7 @@ def load_default_taxonomy() -> None:
         resources = list(map(dict, dict(DEFAULT_TAXONOMY)[resource_type]))
 
         try:
-            upsert_resources(sql_model_map[resource_type], resources)
+            await upsert_resources(sql_model_map[resource_type], resources)
         except QueryError:
             pass  # The upsert_resources function will log the error
         else:
