+++ fidesctl/src/fidesctl/core/apply.py
@@ -87,7 +87,6 @@ def execute_create_update_unchanged(
                 url=url,
                 headers=headers,
                 resource_type=resource_type,
-                resource_id=update_resource.fides_key,
                 json_resource=update_resource.json(exclude_none=True),
             ),
             verbose=False,
